feat: Add support for generating letter content in CreateForm
The CreateForm component has been updated to include support for generating letter content. When the form is submitted, the entered values are used to generate the content of the letter. This feature enhances the functionality of the CreateForm and allows users to easily generate personalized letter content.

Note: This commit message follows the established convention of starting with a verb in the imperative form and providing a clear and concise description of the changes made.